+++ spec/orm/models/person.rb
@@ -10,8 +10,15 @@ class Person < MassiveRecord::ORM::Table
   end
 
   references_one :boss, :class_name => "PersonWithTimestamp", :store_in => :info
+  references_many :test_classes, :store_in => :info
+  references_many :friends, :class_name => "Person", :records_starts_from => :friends_records_starts_from_id
 
   validates_presence_of :name, :age
   validates_numericality_of :age, :greater_than_or_equal_to => 0
   validates_format_of :email, :with => /\A([^@\s]+)@((?:[-a-z0-9]+\.)+[a-z]{2,})\Z/i, :allow_blank => true
+
+
+  def friends_records_starts_from_id
+    id+'-'
+  end
 end