+++ spec/shared/orm/relations/proxy.rb
@@ -1,8 +1,9 @@
 shared_examples_for "relation proxy" do
   let(:target) { mock(Object).as_null_object }
+  let(:find_target_returns) { subject.represents_a_collection? ? [target] : target }
 
   before do
-    subject.metadata = mock(MassiveRecord::ORM::Relations::Metadata).as_null_object if subject.metadata.nil?
+    subject.metadata = mock(MassiveRecord::ORM::Relations::Metadata, :find_with => nil).as_null_object if subject.metadata.nil?
   end
 
   %w(owner target metadata).each do |method|
@@ -48,7 +49,7 @@ shared_examples_for "relation proxy" do
       subject.target = target
       subject.reset
       subject.stub(:find_target?).and_return(false)
-      subject.target.should be_nil
+      subject.target.should be_blank
     end
   end
 
@@ -69,57 +70,18 @@ shared_examples_for "relation proxy" do
 
     it "should return target if loaded successfully" do
       subject.should_receive(:can_find_target?).and_return true
-      subject.should_receive(:find_target) { target }
-      subject.reload.should == target
+      subject.should_receive(:find_target) { find_target_returns }
+      subject.reload.should == find_target_returns
     end
 
     it "should return nil if loading of target failed" do
       subject.stub(:can_find_target?).and_return true
       subject.should_receive(:find_target).and_raise MassiveRecord::ORM::RecordNotFound
-      subject.reload.should be_nil
+      subject.reload.should be_blank
     end
   end
 
 
-  describe "forward method calls to target" do
-    let(:target) { mock(Object, :target_method => "return value", :id => "dummy-id") }
-
-    before do
-      subject.target = target
-    end
-    
-
-    describe "#respond_to?" do
-      it "should check proxy to see if it responds to something" do
-        should respond_to :target
-      end
-      
-      it "should respond to target_method" do
-        should respond_to :target_method
-      end
-
-      it "should not respond to a dummy method" do
-        should_not respond_to :dummy_method_which_does_not_exists 
-      end
-    end
-
-
-    describe "#method_missing" do
-      it "should call proxy's method if exists in proxy" do
-        subject.should_receive(:loaded?).once
-        subject.loaded?
-      end
-
-      it "should call target's method if it responds to it" do
-        target.should_receive(:target_method).and_return(target)
-        subject.target_method.should == target
-      end
-
-      it "should rause no method error if no one responds to it" do
-        lambda { subject.dummy_method_which_does_not_exists }.should raise_error NoMethodError
-      end
-    end
-  end
 
 
   describe "target" do
@@ -136,27 +98,27 @@ shared_examples_for "relation proxy" do
     it "should not try to load target if it has been loaded" do
       subject.loaded!
       should_not_receive :find_target
-      subject.load_target.should be_nil
+      subject.load_target.should be_blank
     end
 
     it "should try to load the target if it has not been loaded" do
       subject.stub(:can_find_target?).and_return true
-      subject.should_receive(:find_target) { target }
+      subject.should_receive(:find_target) { find_target_returns }
       subject.load_target
-      subject.target.should == target
+      subject.target.should == find_target_returns
     end
 
     it "should reset proxy if target's record was not found" do
       subject.stub(:can_find_target?).and_return true
       subject.should_receive(:find_target).and_raise MassiveRecord::ORM::RecordNotFound
-      subject.load_target.should be_nil
+      subject.load_target.should be_blank
     end
   end
 
 
   describe "replace" do
-    let (:old_target) { subject.target_class.new }
-    let (:new_target) { subject.target_class.new }
+    let(:old_target) { subject.represents_a_collection? ? [subject.target_class.new] : subject.target_class.new }
+    let(:new_target) { subject.represents_a_collection? ? [subject.target_class.new] : subject.target_class.new }
 
     before do
       subject.target = old_target
@@ -174,7 +136,7 @@ shared_examples_for "relation proxy" do
   end
 
   describe "find_with" do
-    let(:metadata) { MassiveRecord::ORM::Relations::Metadata.new 'name', :find_with => Proc.new { |target| Person.find("testing-123") }}
+    let(:metadata) { MassiveRecord::ORM::Relations::Metadata.new 'person', :find_with => Proc.new { |target| Person.find("testing-123") }}
     let(:person) { Person.new }
 
     before do